requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.9
pillow==11.1.0
numpy==2.2.3
ultralytics==8.3.64
